Rename pagedriver to browser for clarity
