+++ app/scraper/loot/scraper.py
@@ -9,9 +9,9 @@ from typing import Awaitable, Callable
 
 from playwright.async_api import BrowserContext, Error, Locator, Page
 
+from app.browser import get_new_page
 from app.common import Category, OfferDuration, OfferType, Source
 from app.configparser import Config
-from app.pagedriver import get_new_page
 from app.sqlalchemy import Offer
 
 logger = logging.getLogger(__name__)
