+++ app/scraper/loot/steam_base.py
@@ -5,8 +5,8 @@ from datetime import datetime, timedelta, timezone
 
 from playwright.async_api import Locator, Page
 
+from app.browser import get_new_page
 from app.common import OfferDuration, Source
-from app.pagedriver import get_new_page
 from app.scraper.info.steam import skip_age_verification
 from app.scraper.info.utils import clean_title
 from app.scraper.loot.scraper import OfferHandler, RawOffer, Scraper
